warning message

+++ scripts/buildProtoFiles.sh
@@ -19,6 +19,11 @@ if [ ! -f "$protoc" ]; then
   sh "$scriptsDir"/compileThirdParty.sh
 fi
 
+if [ ! -f "$phpprotoc" ]; then
+  echo "$phpprotoc not found... should be installed from thirdparty bootstrap."
+  exit 1
+fi
+
 mkdir -p "$cppSrcDir"/liblrdpi "$cppSrcDir"/libstats "$cppSrcDir"/libconf $cppSrcDir/libcommand "$cppSrcDir"/libprocess "$cppSrcDir"/liblua "$cppSrcDir"/libtools "$cppSrcDir"/libfork "$cppSrcDir"/liblicense "$cppSrcDir"/libmessages
 
 cp "$protoFileDir"/DpiMsgLRproto.proto "$protoFileDir"/DpiMsgLRproto.proto.orig
